moved idle thread to threads file
